+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.0
+
+Enhancement:
+  - Remove Leaks
+  - Add trusted proxy support
+  - Add node 6.x.x support
+  - Change to MIT licensing
+
 ## 0.0.8
 
 Enhancement: